+++ src/collective/mailchimp/tests/test_setup.py
@@ -3,6 +3,9 @@ from collective.mailchimp.testing import (
     COLLECTIVE_MAILCHIMP_INTEGRATION_TESTING,
 )
 from Products.CMFCore.utils import getToolByName
+from plone import api
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
 
 import unittest
 
@@ -15,6 +18,11 @@ except ImportError:
     PLONE_5 = False
 else:
     PLONE_5 = True
+
+try:
+    from Products.CMFPlone.utils import get_installer
+except ImportError:
+    get_installer = None
 
 
 class TestSetup(unittest.TestCase):
@@ -69,6 +77,35 @@ class TestSetup(unittest.TestCase):
         else:
             # Don't know how to test this in Plone 5.
             pass
+
+
+class TestUninstall(unittest.TestCase):
+
+    layer = COLLECTIVE_MAILCHIMP_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        if get_installer:
+            self.installer = get_installer(self.portal, self.layer['request'])
+        else:
+            self.installer = api.portal.get_tool('portal_quickinstaller')
+        roles_before = api.user.get_roles(TEST_USER_ID)
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.installer.uninstallProducts(['collective.mailchimp'])
+        setRoles(self.portal, TEST_USER_ID, roles_before)
+
+    def test_product_uninstalled(self):
+        """Test if collective.mailchimp is cleanly uninstalled."""
+        self.assertFalse(
+            self.installer.isProductInstalled('collective.mailchimp')
+        )
+
+    def test_browserlayer_removed(self):
+        """Test that ICollectiveMailchimp is removed."""
+        from collective.mailchimp.interfaces import ICollectiveMailchimp
+        from plone.browserlayer import utils
+
+        self.assertNotIn(ICollectiveMailchimp, utils.registered_layers())
 
 
 def test_suite():